feat: add SpRegionChip, use VerificationChip in variants, increase header gap
- SpRegionChip: new Role component with flag emoji and filled styling
- Game detail header: increased gap between title and badges (Medium → Default)
- Variants section: replaced plain SpChip with VerificationChip for verification status
- Variants section: replaced inline region chip with SpRegionChip
- Removed duplicated regionFlags map from GameDetailScreen
